settings/tokens: Extract `index` route
This change is in preparation for having a `settings/tokens/new` route soon
