+++ app/templates/settings/tokens.hbs
@@ -3,5 +3,5 @@
 <PageHeader @title="Account Settings" />
 
 <SettingsPage>
-  <Settings::ApiTokens @tokens={{@model}} />
+  {{outlet}}
 </SettingsPage>
