+++ app/templates/settings/tokens/index.hbs
@@ -1,0 +1,1 @@
+<Settings::ApiTokens @tokens={{@model}} />
